generalized branin dataset to grid dataset; generalized oracle filter to score filter; created oracle mappers; testing of different settings; init hartmann
