+++ my_package/config/filter/score.yaml
@@ -1,0 +1,1 @@
+_target_: filter.filter.ScoreFilter
