--- my_package/config/logger/wandb.yaml
@@ -1,51 +1,0 @@
-_target_: gflownet.utils.logger.Logger
-
-do:
-  online: True
-  times: False
-
-project_name: "ActiveLearningLoop_Test"
-
-# Train metrics
-train:
-  period: 1
-# Test metrics
-test:
-  first_it: True
-  period: 100
-  n: 100
-  kde:
-    bandwidth: 0.1
-    kernel: gaussian
-  n_top_k: 5000
-  top_k: 100
-  top_k_period: -1
-  # Number of backward trajectories to estimate the log likelihood of each test data point
-  n_trajs_logprobs: 10
-  logprobs_batch_size: 100
-  logprobs_bootstrap_size: 10000
-  # Maximum number of test data points to compute log likelihood probs.
-  max_data_logprobs: 1e5
-# Oracle metrics
-oracle:
-  period: 100000
-  k:
-    - 1
-    - 10
-    - 100
-# Policy model checkpoints
-checkpoints:
-  period: 1000
-
-# Log dir
-logdir:
-  root: ./logs
-  ckpts: ckpts
-  overwrite: True
-debug: False
-lightweight: False
-progress: True
-context: "0"
-notes: null # wandb run notes (e.g. "baseline")
-tags: 
-    - gflownet
