+++ my_package/config/oracle/multifidelity.yaml
@@ -1,0 +1,8 @@
+_target_: oracle.oracle.MultiFidelityOracle
+
+oracles:
+  - branin1:
+      _target_: gflownet.proxy.box.branin.Branin
+      fidelity: 1
+      do_domain_map: True
+      cost: 100
