+++ storage/hartmann/mf/data_train.csv
@@ -1,3 +1,126 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3b209a3813c3d36aa88ff795883f742b6fc719493470a9d55a4b0a7bdb93254f
-size 2007
+samples
+[4 5 7 8 6 6];1
+[0 8 8 7 0 0];0
+[5 0 0 5 2 2];1
+[5 0 7 7 9 2];0
+[8 0 9 5 9 8];1
+[0 6 7 0 0 0];0
+[2 8 1 9 7 9];0
+[6 9 6 7 0 6];0
+[5 1 4 1 2 1];1
+[2 5 6 6 8 6];0
+[6 9 1 1 3 9];1
+[3 5 1 5 8 9];2
+[4 1 7 2 9 7];1
+[4 7 8 6 4 5];0
+[6 9 5 9 0 2];0
+[7 5 3 2 2 9];0
+[4 4 1 6 8 1];1
+[2 5 3 5 2 1];0
+[4 9 8 9 7 3];0
+[3 7 3 1 1 7];0
+[0 1 1 3 0 6];1
+[4 8 6 5 1 5];1
+[7 6 0 2 2 6];1
+[6 3 1 3 1 4];0
+[1 4 1 7 3 7];0
+[4 4 1 6 8 1];1
+[4 5 5 5 3 3];0
+[5 7 7 6 2 3];0
+[3 3 6 1 6 8];0
+[0 2 6 3 0 6];0
+[3 5 5 6 6 2];1
+[7 8 6 6 9 0];0
+[6 6 7 3 4 6];1
+[3 8 6 3 8 2];0
+[3 6 0 7 9 2];1
+[1 4 6 0 6 4];0
+[4 7 4 2 2 5];0
+[2 0 5 5 6 5];1
+[0 6 5 5 4 5];1
+[5 0 4 2 1 7];0
+[2 0 4 1 6 6];1
+[5 9 1 3 8 1];1
+[5 2 3 1 5 7];0
+[2 1 0 3 4 4];0
+[2 5 9 5 0 9];1
+[3 7 4 8 5 1];0
+[6 9 7 6 0 2];0
+[4 5 1 4 4 6];1
+[1 1 6 4 5 9];0
+[1 5 6 0 2 8];0
+[7 3 6 2 2 6];0
+[4 3 5 4 5 5];1
+[0 4 5 3 5 6];0
+[5 8 2 8 1 1];1
+[2 8 5 5 0 1];0
+[6 0 3 4 4 8];0
+[2 9 7 7 3 0];1
+[4 7 8 4 5 2];0
+[2 3 9 1 3 8];1
+[4 6 9 6 3 2];0
+[2 3 8 5 2 7];0
+[2 9 8 7 7 0];0
+[2 3 8 5 2 7];0
+[1 3 1 3 4 9];0
+[4 2 5 5 2 5];1
+[4 7 4 5 7 2];0
+[3 3 7 5 3 6];0
+[5 9 7 8 0 1];0
+[2 0 6 5 4 6];0
+[4 8 4 8 5 1];0
+[5 3 5 2 2 8];0
+[1 2 3 5 4 6];0
+[1 3 4 1 4 5];0
+[2 5 7 2 2 8];0
+[3 4 6 5 3 6];0
+[1 1 4 4 3 4];0
+[3 2 7 4 2 9];0
+[1 3 4 1 4 5];1
+[4 9 8 7 3 2];0
+[4 9 5 8 2 0];1
+[3 1 4 1 3 8];0
+[0 3 3 3 2 7];0
+[0 1 7 2 4 7];2
+[1 0 4 1 4 6];1
+[4 9 5 8 2 0];0
+[3 2 1 4 2 6];0
+[4 9 5 8 2 0];0
+[2 0 6 1 2 7];0
+[2 5 4 2 3 6];0
+[5 2 6 3 4 6];2
+[3 8 8 7 5 0];0
+[3 9 9 5 7 0];0
+[3 4 2 3 3 6];0
+[5 8 0 6 0 1];2
+[4 9 5 4 7 1];1
+[4 7 8 6 9 0];1
+[3 3 5 3 4 8];0
+[3 4 2 3 3 6];0
+[5 8 0 6 8 1];0
+[2 0 3 3 3 5];0
+[4 8 8 7 8 1];0
+[4 8 1 7 2 1];0
+[1 0 4 1 3 7];0
+[4 8 8 7 8 1];1
+[5 8 7 6 4 1];1
+[4 9 5 7 0 1];0
+[4 9 8 7 2 0];0
+[5 9 7 6 4 1];1
+[3 8 5 6 0 0];1
+[3 9 9 6 5 1];1
+[1 2 4 2 4 7];0
+[4 9 8 7 2 0];1
+[3 8 5 6 0 0];0
+[1 2 4 2 4 7];0
+[1 2 5 4 3 6];1
+[2 3 3 3 3 8];0
+[2 3 3 3 3 8];0
+[4 9 9 5 1 1];1
+[4 9 3 5 2 1];1
+[4 9 9 5 1 1];1
+[4 9 4 5 4 1];0
+[4 9 3 5 2 1];0
+[4 9 0 5 2 1];0
+[4 9 0 5 2 1];2
+[4 9 0 5 2 1];0
